+++ localization/vision_common/CMakeLists.txt
@@ -75,7 +75,6 @@ if(CATKIN_ENABLE_TESTING)
   add_rostest_gtest(test_ransac_pnp
     test/test_ransac_pnp.test
     test/test_ransac_pnp.cc
-    test/test_utilities.cc
   )
   target_link_libraries(test_ransac_pnp
     ${PROJECT_NAME} ${catkin_LIBRARIES} 
@@ -83,7 +82,6 @@ if(CATKIN_ENABLE_TESTING)
   add_rostest_gtest(test_reprojection_pose_estimate
     test/test_reprojection_pose_estimate.test
     test/test_reprojection_pose_estimate.cc
-    test/test_utilities.cc
   )
   target_link_libraries(test_reprojection_pose_estimate
     ${PROJECT_NAME} ${catkin_LIBRARIES} 
@@ -91,7 +89,6 @@ if(CATKIN_ENABLE_TESTING)
   add_rostest_gtest(test_pnp
     test/test_pnp.test
     test/test_pnp.cc
-    test/test_utilities.cc
   )
   target_link_libraries(test_pnp
     ${PROJECT_NAME} ${catkin_LIBRARIES} 
@@ -99,7 +96,6 @@ if(CATKIN_ENABLE_TESTING)
   add_rostest_gtest(test_camera_utilities
     test/test_camera_utilities.test
     test/test_camera_utilities.cc
-    test/test_utilities.cc
   )
   target_link_libraries(test_camera_utilities
     ${PROJECT_NAME} ${catkin_LIBRARIES}
