+++ localization/vision_common/test/test_pnp.cc
@@ -16,12 +16,11 @@
  * under the License.
  */
 
-#include "test_utilities.h"  // NOLINT
-
 #include <localization_common/logger.h>
 #include <localization_common/test_utilities.h>
 #include <vision_common/camera_utilities.h>
 #include <vision_common/identity_distorter.h>
+#include <vision_common/test_utilities.h>
 
 #include <opencv2/calib3d/calib3d.hpp>
